requirements.txt
# ShadowRecon v1.0 - Ultimate Web Attack Surface Discovery Framework
# Developed by kernelpanic | A product of infosbios

# Core HTTP and networking
aiohttp>=3.8.0
aiofiles>=0.8.0
dnspython>=2.2.0
cryptography>=3.4.0

# Web parsing and processing
beautifulsoup4>=4.10.0
lxml>=4.6.0
requests>=2.28.0

# Data processing and utilities  
colorama>=0.4.4
tqdm>=4.62.0
python-dateutil>=2.8.0

# Optional dependencies for advanced features
# Install with: pip install shadowrecon[full]

# For SSL/TLS analysis
pyOpenSSL>=22.0.0

# For advanced DNS analysis
dns-lexicon>=3.8.0

# For screenshot capture (optional)
selenium>=4.0.0; extra == "screenshots"
webdriver-manager>=3.8.0; extra == "screenshots"

# For database storage (optional) 
sqlalchemy>=1.4.0; extra == "database"
sqlite3; extra == "database"

# Development dependencies (install with: pip install shadowrecon[dev])
pytest>=6.2.0; extra == "dev"
pytest-asyncio>=0.18.0; extra == "dev"
pytest-cov>=2.12.0; extra == "dev"
black>=22.0.0; extra == "dev"
flake8>=4.0.0; extra == "dev"
mypy>=0.950; extra == "dev"
bandit>=1.7.0; extra == "dev"
safety>=2.0.0; extra == "dev"
pre-commit>=2.15.0; extra == "dev"

# Documentation (install with: pip install shadowrecon[docs])
sphinx>=4.5.0; extra == "docs"
sphinx-rtd-theme>=1.0.0; extra == "docs"
